+++ src/components/ui/progress/Progress.tsx
@@ -4,6 +4,8 @@ import { cap, scale } from "@ezez/utils";
 
 import type { ThemeCSS } from "../../../theme";
 
+import { fnWithProps } from "../../../types/fnWithProps";
+
 import { Background, Container, Value } from "./Progress.styled";
 
 interface Props extends React.ComponentProps<typeof Container> {
@@ -25,12 +27,19 @@ interface Props extends React.ComponentProps<typeof Container> {
     scaleTo?: number;
 }
 
+type Selectors = {
+    selectors: {
+        background: string;
+        value: string;
+    };
+};
+
 /**
  * Progress bar component, it can be filled from the start or from any arbitrary point.
  *
  * At the moment, no value (undefined) means starting point, but in the future it will mean "undetermined" state.
  */
-const Progress: React.FC<Props> = (props) => {
+const Progress = fnWithProps<React.FC<Props>, Selectors>((props) => {
     // eslint-disable-next-line @typescript-eslint/no-magic-numbers
     const { value, valueFrom, scaleFrom = 0, scaleTo = 100, ...rest } = props;
 
@@ -59,10 +68,12 @@ const Progress: React.FC<Props> = (props) => {
             <Value css={css} zero={isZeroWidth} />
         </Container>
     );
-};
+}, {
+    selectors: {
+        background: Background.toString(),
+        value: Value.toString(),
+    },
+});
 
-const BackgroundClassName = Background.toString();
-const ValueClassName = Value.toString();
-
-export { Progress, BackgroundClassName, ValueClassName };
+export { Progress };